removed duplicates of the ProtocolIDs
