+++ apps/dashboard_app/app/schemas/__init__.py
@@ -1,4 +1,4 @@
-from .watcher import NotificationForm
-from .auth import *
-from .loan_state import *
-from .user import *
+from schemas.watcher import NotificationForm
+from schemas.auth import *
+from schemas.loan_state import *
+from schemas.user import *
