+++ apps/dashboard_app/app/telegram_app/telegram/handlers/__init__.py
@@ -1,8 +1,8 @@
 from aiogram import Router
 
-from .command import cmd_router
-from .menu import menu_router
-from .create_notification import create_notification_router
+from handlers.command import cmd_router
+from handlers.menu import menu_router
+from handlers.create_notification import create_notification_router
 
 # Create the main router to simplify imports
 index_router = Router()
