+++ apps/dashboard_app/app/utils/values.py
@@ -1,6 +1,6 @@
 import os
 from dataclasses import dataclass
-from enum import Enum
+from shared.protocol_ids import ProtocolIDs
 
 from dotenv import load_dotenv
 
@@ -10,12 +10,7 @@ DATA_HANDLER_ENDPOINT = os.environ.get("DATA_HANDLER_URL", "")
 HEALTH_RATIO_URL = (
     f"{DATA_HANDLER_ENDPOINT}/health-ratio-per-user/{{protocol}}/?user_id={{user_id}}"
 )
-CURRENTLY_AVAILABLE_PROTOCOL_IDS: tuple[str, ...] = (
-    "zkLend",
-    "Nostra_alpha",
-    "Nostra_mainnet",
-    "Vesu",
-)
+CURRENTLY_AVAILABLE_PROTOCOL_IDS = tuple(item.value for item in ProtocolIDs)
 
 
 @dataclass(frozen=True)
@@ -42,13 +37,6 @@ class CreateSubscriptionValues:
     create_subscription_description_message: str = (
         "Creates a new subscription to notifications"
     )
-
-
-class ProtocolIDs(Enum):
-    HASHSTACK = "Hashstack"
-    NOSTRA_ALPHA = "Nostra_alpha"
-    NOSTRA_MAINNET = "Nostra_mainnet"
-    ZKLEND = "zkLend"
 
 
 HEALTH_RATIO_LEVEL_ALERT_VALUE: float = 0.1
