+++ apps/data_handler/tests/order_book/test_myswap.py
@@ -7,7 +7,7 @@ from unittest.mock import AsyncMock, Mock, patch
 
 import pytest
 
-from ...handlers.order_books.myswap.main import MySwapOrderBook
+from data_handler.handlers.order_books.myswap.main import MySwapOrderBook
 
 MAX_MYSWAP_TICK = Decimal("1774532")
