+++ apps/shared/data_parser/zklend.py
@@ -3,7 +3,7 @@ This module contains the logic to parse the zkLend data to human-readable format
 """
 
 from typing import Any, List
-from .serializers import (
+from data_parser.serializers import (
     AccumulatorsSyncEventData,
     LiquidationEventData,
     WithdrawalEventData,
