+++ apps/shared/error_handler/notifications.py
@@ -4,8 +4,8 @@ from uuid import uuid4
 
 from aiogram import Bot, Dispatcher
 
-from .config import ERROR_CHAT_ID, TELEGRAM_TOKEN
-from .values import Message
+from error_handler.config import ERROR_CHAT_ID, TELEGRAM_TOKEN
+from error_handler.values import Message
 
 logger = logging.getLogger(__name__)
 logging.basicConfig(level=logging.INFO)
