+++ apps/shared/loan_entity/nostra/alpha/entity.py
@@ -3,7 +3,7 @@ Event processing for Nostra Alpha protocol, including loan and collateral manage
 """
 
 import decimal
-from .settings import NOSTRA_ALPHA_SPECIFIC_TOKEN_SETTINGS
+from alpha.settings import NOSTRA_ALPHA_SPECIFIC_TOKEN_SETTINGS
 from shared.custom_types import InterestRateModels, Portfolio, Prices, TokenParameters
 from shared.loan_entity import LoanEntity
