+++ apps/shared/loan_entity/nostra/mainnet/__init__.py
@@ -1,2 +1,2 @@
-from .entity import NostraMainnetLoanEntity
-from .settings import *
+from mainnet.entity import NostraMainnetLoanEntity
+from mainnet.settings import *
