+++ apps/shared/tests/test_helpers.py
@@ -1,5 +1,5 @@
-from ..helpers import add_leading_zeros, get_addresses, get_symbol
-from ..custom_types.base import TokenParameters
+from dashboard_app.helpers import add_leading_zeros, get_addresses, get_symbol
+from shared.custom_types.base import TokenParameters
 from unittest.mock import AsyncMock, patch
